Rename reducers to index
