+++ app/store/configureStore.dev.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension'
-import rootReducer from './reducers'
+import rootReducer from '../modules'
 import thunk from 'redux-thunk'
 import { createLogger } from 'redux-logger'
