+++ app/store/configureStore.prod.js
@@ -1,5 +1,5 @@
 import { createStore, applyMiddleware } from 'redux'
-import rootReducer from './reducers'
+import rootReducer from '../modules'
 import thunk from 'redux-thunk'
 
 function configureStore (initialState) {
